refactor(api): Split handlers into focused modules
